wip refactor snp frequencies
