+++ malariagen_data/anoph/genome_features.py
@@ -1,4 +1,4 @@
-from typing import Dict, Optional, Tuple
+from typing import Dict, Optional, Tuple, Mapping
 
 import bokeh.models
 import bokeh.plotting
@@ -26,6 +26,7 @@ class AnophelesGenomeFeaturesData(AnophelesGenomeSequenceData):
         *,
         gff_gene_type: str,
         gff_default_attributes: Tuple[str, ...],
+        gene_names: Optional[Mapping[str, str]] = None,
         **kwargs,
     ):
         # N.B., this class is designed to work cooperatively, and
@@ -37,6 +38,11 @@ class AnophelesGenomeFeaturesData(AnophelesGenomeSequenceData):
         # change if the GFF ever changed.
         self._gff_gene_type = gff_gene_type
         self._gff_default_attributes = gff_default_attributes
+
+        # Allow manual override of gene names.
+        if gene_names is None:
+            gene_names = dict()
+        self._gene_names = gene_names
 
         # Setup caches.
         self._cache_genome_features: Dict[Tuple[str, ...], pd.DataFrame] = dict()
@@ -429,3 +435,18 @@ class AnophelesGenomeFeaturesData(AnophelesGenomeSequenceData):
         fig.xaxis.ticker = bokeh.models.AdaptiveTicker(min_interval=1)
         fig.xaxis.minor_tick_line_color = None
         fig.xaxis[0].formatter = bokeh.models.NumeralTickFormatter(format="0,0")
+
+    def _transcript_to_gene_name(self, transcript):
+        df_genome_features = self.genome_features().set_index("ID")
+        rec_transcript = df_genome_features.loc[transcript]
+        parent_id = rec_transcript["Parent"]
+
+        if parent_id in self._gene_names:
+            # Manual override.
+            parent_name = self._gene_names[parent_id]
+        else:
+            rec_parent = df_genome_features.loc[parent_id]
+            # Try to access "Name", fall back to ID if not present.
+            parent_name = rec_parent.get("Name", parent_id)
+
+        return parent_name
